# openapi/gpt-actions.yaml
openapi: 3.1.0
info:
  title: Persistent GPT Memory MCP
  version: 0.1.0
  description: REST API for a private local-file persistent memory used by GPT Actions.
servers:
  - url: https://{domain}
    description: Production HTTPS endpoint
    variables:
      domain:
        default: memory.example.com
        description: Replace with APP_DOMAIN.
  - url: http://localhost:8787
    description: Local development endpoint
security: []
paths:
  /health:
    get:
      operationId: getMemoryHealth
      summary: Check memory server health.
      description: Use this to verify that the memory service, storage path and index are available.
      responses:
        "200":
          description: Health status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /api/index:
    get:
      operationId: listMemoryIndex
      summary: List compact memory index.
      description: Use before reading documents. Returns metadata and summaries only.
      parameters:
        - $ref: "#/components/parameters/Type"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/Tag"
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Compact document index.
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: "#/components/schemas/IndexedDocument"
                required: [documents]
  /api/docs:
    get:
      operationId: listMemoryDocuments
      summary: List memory documents by metadata.
      description: Lists document summaries. Does not return full Markdown content.
      parameters:
        - $ref: "#/components/parameters/Type"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/Tag"
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Matching document summaries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: "#/components/schemas/IndexedDocument"
                required: [documents]
    post:
      operationId: createMemoryDocument
      summary: Create a memory document.
      description: Create a short reusable Markdown memory document after checking for duplicates with search.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentInput"
      responses:
        "201":
          description: Created document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryDocument"
  /api/docs/{id}:
    get:
      operationId: readMemoryDocument
      summary: Read one memory document.
      description: Read full Markdown content only after selecting the document from the index or search.
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Full document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryDocument"
    put:
      operationId: updateMemoryDocument
      summary: Update one memory document.
      description: Updates content and selected metadata. Requires expected_version and archives the previous version automatically.
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentInput"
      responses:
        "200":
          description: Updated document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryDocument"
    delete:
      operationId: deleteMemoryDocument
      summary: Move one document to trash.
      description: Does not permanently delete. Requires a reason and moves the document to trash.
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteDocumentInput"
      responses:
        "200":
          description: Deleted document summary in trash.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndexedDocument"
  /api/docs/{id}/restore:
    post:
      operationId: restoreMemoryDocument
      summary: Restore a trashed document.
      description: Restores a document from trash and updates the index.
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreDocumentInput"
      responses:
        "200":
          description: Restored document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryDocument"
  /api/docs/{id}/permanent:
    delete:
      operationId: permanentlyDeleteMemoryDocument
      summary: Permanently delete a document.
      description: Destructive endpoint. Use only when explicitly requested. Requires confirm=PERMANENT_DELETE and a reason.
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermanentDeleteDocumentInput"
      responses:
        "200":
          description: Permanent delete confirmation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  deleted:
                    type: boolean
                    const: true
                required: [id, deleted]
  /api/docs/{id}/versions:
    get:
      operationId: listDocumentVersions
      summary: List archived versions.
      description: Lists previous versions saved before updates or version restores.
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Version list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  versions:
                    type: array
                    items:
                      $ref: "#/components/schemas/DocumentVersionInfo"
                required: [versions]
  /api/docs/{id}/versions/{version}:
    get:
      operationId: readDocumentVersion
      summary: Read an archived version.
      description: Returns the full Markdown and metadata for one archived version.
      parameters:
        - $ref: "#/components/parameters/DocumentId"
        - $ref: "#/components/parameters/Version"
      responses:
        "200":
          description: Archived document version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryDocument"
  /api/docs/{id}/versions/{version}/restore:
    post:
      operationId: restoreDocumentVersion
      summary: Restore an archived version.
      description: Restores an old version as a new current version while preserving the previous current version.
      parameters:
        - $ref: "#/components/parameters/DocumentId"
        - $ref: "#/components/parameters/Version"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Restored document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryDocument"
  /api/docs/{id}/versions/{version}/diff:
    get:
      operationId: compareDocumentVersion
      summary: Compare archived version with current.
      description: Returns a simple line-by-line diff between an archived version and the current document.
      parameters:
        - $ref: "#/components/parameters/DocumentId"
        - $ref: "#/components/parameters/Version"
      responses:
        "200":
          description: Diff result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  document_id:
                    type: string
                  compared_version:
                    type: integer
                  current_version:
                    type: integer
                  diff:
                    type: string
                required: [document_id, compared_version, current_version, diff]
  /api/search:
    post:
      operationId: searchMemory
      summary: Search memory documents.
      description: Search the compact index by term, type and tags before opening documents.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchInput"
      responses:
        "200":
          description: Search results.
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  results:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/IndexedDocument"
                        - type: object
                          properties:
                            score:
                              type: number
                          required: [score]
                required: [query, results]
  /api/context-pack:
    post:
      operationId: createContextPack
      summary: Create a compact context pack.
      description: Selects relevant documents and combines them into a bounded Markdown context pack for a conversation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContextPackInput"
      responses:
        "200":
          description: Context pack.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContextPack"
  /openapi.yaml:
    get:
      operationId: getOpenApiYaml
      summary: Download the GPT Actions OpenAPI YAML.
      responses:
        "200":
          description: OpenAPI YAML.
          content:
            application/yaml:
              schema:
                type: string
components:
  parameters:
    DocumentId:
      name: id
      in: path
      required: true
      schema:
        type: string
    Version:
      name: version
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    Type:
      name: type
      in: query
      schema:
        $ref: "#/components/schemas/DocumentType"
    Category:
      name: category
      in: query
      schema:
        type: string
    Tag:
      name: tag
      in: query
      schema:
        type: string
    Status:
      name: status
      in: query
      schema:
        $ref: "#/components/schemas/DocumentStatus"
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
  schemas:
    DocumentType:
      type: string
      enum: [context, skill, prompt, process, template, project, decision, reference, workflow, correction_criteria, class_plan]
    DocumentStatus:
      type: string
      enum: [active, archived, deleted]
    DocumentPriority:
      type: string
      enum: [low, medium, high]
    DocumentMetadata:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        type:
          $ref: "#/components/schemas/DocumentType"
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/DocumentStatus"
        priority:
          $ref: "#/components/schemas/DocumentPriority"
        created_at:
          type: string
        updated_at:
          type: string
        version:
          type: integer
          minimum: 1
        summary:
          type: string
      required: [id, slug, title, type, category, tags, status, priority, created_at, updated_at, version, summary]
    IndexedDocument:
      allOf:
        - $ref: "#/components/schemas/DocumentMetadata"
        - type: object
          properties:
            path:
              type: string
          required: [path]
    MemoryDocument:
      type: object
      properties:
        metadata:
          $ref: "#/components/schemas/DocumentMetadata"
        content:
          type: string
        path:
          type: string
      required: [metadata, content, path]
    CreateDocumentInput:
      type: object
      properties:
        title:
          type: string
        type:
          $ref: "#/components/schemas/DocumentType"
        category:
          type: string
          default: general
        tags:
          type: array
          items:
            type: string
        summary:
          type: string
        content:
          type: string
          description: Markdown content.
        priority:
          $ref: "#/components/schemas/DocumentPriority"
      required: [title, type, category, tags, summary, content, priority]
    UpdateDocumentInput:
      type: object
      properties:
        content:
          type: string
        summary:
          type: string
        tags:
          type: array
          items:
            type: string
        expected_version:
          type: integer
          minimum: 1
        change_summary:
          type: string
      required: [content, expected_version, change_summary]
    DeleteDocumentInput:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
    RestoreDocumentInput:
      type: object
      properties:
        reason:
          type: string
    PermanentDeleteDocumentInput:
      type: object
      properties:
        confirm:
          type: string
          const: PERMANENT_DELETE
        reason:
          type: string
      required: [confirm, reason]
    SearchInput:
      type: object
      properties:
        query:
          type: string
        types:
          type: array
          items:
            $ref: "#/components/schemas/DocumentType"
        tags:
          type: array
          items:
            type: string
        limit:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
      required: [query]
    ContextPackInput:
      type: object
      properties:
        goal:
          type: string
        query:
          type: string
        max_documents:
          type: integer
          minimum: 1
          maximum: 20
          default: 5
        max_chars:
          type: integer
          minimum: 1000
          maximum: 60000
          default: 12000
      required: [goal, query]
    ContextPack:
      type: object
      properties:
        goal:
          type: string
        selected_documents:
          type: array
          items:
            $ref: "#/components/schemas/IndexedDocument"
        context_markdown:
          type: string
      required: [goal, selected_documents, context_markdown]
    DocumentVersionInfo:
      type: object
      properties:
        document_id:
          type: string
        version:
          type: integer
        path:
          type: string
        created_at:
          type: string
        summary:
          type: string
      required: [document_id, version, path, created_at]
    Health:
      type: object
      properties:
        status:
          type: string
          const: ok
        storage_path:
          type: string
        document_count:
          type: integer
        active_count:
          type: integer
        deleted_count:
          type: integer
        index_updated_at:
          type: string
        max_document_chars:
          type: integer
      required: [status, storage_path, document_count, active_count, deleted_count, index_updated_at, max_document_chars]